% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractrows.R
\name{extractrows}
\alias{extractrows}
\title{Extract a sub-set of rows (such as a group of gene ids) from a data table. All columns will be preserved.}
\usage{
extractrows(fnamein, fids, fnameout, wdir)
}
\arguments{
\item{fnamein}{character vector that specifies the name of tab-delimited text file with the input data table.}

\item{fids}{character vector that specifies the name of text file with row ids (such as gene ids).
The file must contain one row id per line, without any trailing spaces or any other additional symbols.}

\item{fnameout}{character vector that specifies the name of output file where the new data table will be saved.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.}
}
\value{
no return value
}
\description{
Extract a sub-set of rows (such as a group of gene ids) from a data table. All columns will be preserved.
}
\examples{

Basic usage:
extractrows(fnamein="table.txt", fids="geneids.txt", fnameout="newtable.txt", wdir="c:/test")

Example with built-in data files:

library(data.table)
library(kmcut)

# Load example gene expression data table for 2 genes
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")
# Load a list that contains one gene id (MYCN)
idlist = system.file("extdata", "rowids.txt", package="kmcut")
# Run the function
extractrows(fnamein=fdat, fids=idlist, fnameout="example_genes_subset.txt", wdir="c:/test")

This will create a tab-delimited text file with one row "MYCN" in directory "c:/test":
"example_genes_subset.txt"
}
