% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmoptscut.R
\name{kmoptscut}
\alias{kmoptscut}
\title{For each feature finds a cutoff that optimally stratifies samples into 2 groups, plots Kaplan-Meier survival curves for these groups,
and observed vs. expected optimization plot. Does not use the permutation test to estimate the statistical significance of the optimal cutoff.}
\usage{
kmoptscut(fname, sfname, wdir, min_fraction = 0.1, min_up_down = 1,
  peak_tolerance = 0.1, min_uval = 50, psort = FALSE,
  wlabels = TRUE, wpdf = TRUE, verbose = TRUE)
}
\arguments{
\item{fname}{character vector that specifies the name of the file with feature(s) for each sample. The file must be tab-delimited,
where features are in rows and samples are in columns. First column must contain feature names. Column names must contain sample ids.}

\item{sfname}{character vector that specifies the name of the file with right-censored survival time data. The file must be tab-delimited,
where samples are in rows. First column must contain sample ids that match those in 'fname'. The file must contain columns called 'stime' and 'scens',
with survival time and censoring variable (0 or 1), respectively.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.
Output file names are automatically created by adding\cr"_KMopt_minf_.2f" and corresponding extension to 'fname'.}

\item{min_fraction}{numeric value that specifies the minimal fraction of samples in the smaller group (default is 0.1).}

\item{min_up_down}{numeric value that specifies the minimal number of up/down points on either side of the peak for
'pracma::findpeaks' function (default is 1).}

\item{peak_tolerance}{numeric value that specifies the maximal difference between in heigth between top peaks.
The peak within 'peak_tolerance' closest to the median value is selected.}

\item{min_uval}{numeric value that specifies the minimal percentage of unique values per feature (default is 50).
Features that have less than 'min_uval' percent unique values are excluded from the analysis.}

\item{psort}{logical value whether to sort the output table by p-values in increasing order (default is FALSE).}

\item{wlabels}{logical value whether to write a CSV file with low/high (below/above the cutoff) group sample labels (default is TRUE).}

\item{wpdf}{logical value whether to write a PDF file with survival curves and plots (default is TRUE).}

\item{verbose}{logical value whether to print progress (default is TRUE).}
}
\value{
no return value
}
\description{
For each feature finds a cutoff that optimally stratifies samples into 2 groups, plots Kaplan-Meier survival curves for these groups,
and observed vs. expected optimization plot. Does not use the permutation test to estimate the statistical significance of the optimal cutoff.
}
\examples{
Basic usage:
kmoptscut(fname="table.txt", sfname="survival.txt", wdir="c:/test")

Example with built-in data files:

library(stats)
library(survival)
library(stringr)
library(data.table)
library(tools)
library(pracma)
library(kmcut)

# Load example gene expression data and survival data for 2 genes and 295 samples
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")
sdat = system.file("extdata", "survival_data_295.txt", package="kmcut")

kmoptscut(fname=fdat, sfname=sdat, wdir="c:/test")

This will create three output files in directory "c:/test":
1) PDF file with plots: "example_genes_295_KMopt_minf_0.10.pdf"
2) Tab-delimited text file with the results: "example_genes_295_KMopt_minf_0.10.txt"
3) CSV file with low/high sample labels: "example_genes_295_KMopt_minf_0.10_labels.csv"
}
