% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmucut.R
\name{kmucut}
\alias{kmucut}
\title{For each feature uses the user-supplied cutoff to stratify samples into 2 groups,
plots Kaplan-Meier survival curves, and calculates the log-rank test p-value}
\usage{
kmucut(fname, sfname, wdir, cutoff, min_uval = 50, psort = FALSE,
  wlabels = TRUE, wpdf = TRUE)
}
\arguments{
\item{fname}{character vector that specifies the name of the file with feature(s) for each sample. The file must be tab-delimited,
where features are in rows and samples are in columns. First column must contain feature names. Column names must contain sample ids.}

\item{sfname}{character vector that specifies the name of the file with right-censored survival time data. The file must be tab-delimited,
where samples are in rows. First column must be named 'sample_id' and contain sample ids that match those in 'fname'. The file must contain columns called 'stime' and 'scens',
with survival time and censoring variable (0 or 1), respectively.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.
Output file names are automatically created by adding\cr"_KM_ucut_.2f" and corresponding extension to 'fname'.}

\item{cutoff}{numeric value that specifies the cutoff value for stratification.The same cutoff is applied to every feature in the dataset.}

\item{min_uval}{numeric value that specifies the minimal percentage of unique values per feature (default is 50).
Features that have less than 'min_uval' percent unique values are excluded from the analysis.}

\item{psort}{logical value whether to sort the output table by p-values in increasing order (default is FALSE).}

\item{wlabels}{logical value whether to write a CSV file with low/high (below/above the cutoff) group sample labels (default is TRUE).}

\item{wpdf}{logical value whether to write a PDF file with survival curves and plots (default is TRUE).}
}
\value{
no return value
}
\description{
For each feature uses the user-supplied cutoff to stratify samples into 2 groups,
plots Kaplan-Meier survival curves, and calculates the log-rank test p-value
}
\examples{

Basic usage:
kmucut(fname="table.txt", sfname="survival.txt", wdir="c:/test", cutoff=2.5)

Example with built-in data files:

library(stats)
library(survival)
library(stringr)
library(data.table)
library(tools)
library(pracma)
library(kmcut)

# Load example gene expression data and survival data for 2 genes and 295 samples
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")
sdat = system.file("extdata", "survival_data_295.txt", package="kmcut")

kmucut(fname=fdat, sfname=sdat, wdir="c:/test", cutoff=5, min_uval=90)

This will create three output files in directory "c:/test":
1) PDF file with plots: "example_genes_295_KM_ucut_5.pdf"
2) Tab-delimited text file with the results: "example_genes_295_KM_ucut_5.txt"
3) CSV file with low/high sample labels: "example_genes_295_KM_ucut_5_labels.csv"
}
