% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmvalcut.R
\name{kmvalcut}
\alias{kmvalcut}
\title{Creates Kaplan-Meier survival curves for each feature from a validation data set by using a file with previously determined
stratification thresholds (one threshold per feature), and calculates the log-rank test p-value.}
\usage{
kmvalcut(input1, input2, sfname, wdir, min_uval = 50, psort = FALSE,
  wlabels = TRUE, wpdf = TRUE)
}
\arguments{
\item{input1}{character vector that specifies the name of tab-delimited file with the table that contains one or more feature
and a stratification threshold for each feature (this table is produced by kmoptscut, kmoptpermcut, kmqcut or kmucut).
The file with previously determined stratification thresholds must have first two columns named as 'tracking_id' and 'CUTOFF'.
The 'tracking_id' column contains feature names, the 'CUTOFF' column contains stratification threshold for each feature.}

\item{input2}{character vector that specifies the name of the file with feature(s) for each sample. The file must be tab-delimited,
where features are in rows and samples are in columns. First column must contain feature names. Column names must contain sample ids.
Feature names must exactly match the ones in 'input1' file.}

\item{sfname}{character vector that specifies the name of the file with right-censored survival time data. The file must be tab-delimited,
where samples are in rows. First column must be named 'sample_id' and contain sample ids that match those in 'fname'. The file must contain columns called 'stime' and 'scens',
with survival time and censoring variable (0 or 1), respectively.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.
Output file names are automatically created by adding\cr'_KM_val' and corresponding extension to 'input2'.}

\item{min_uval}{numeric value that specifies the minimal percentage of unique values per feature (default is 50).
Features that have less than 'min_uval' percent unique values are excluded from the analysis.}

\item{psort}{logical value whether to sort the output table by p-values in increasing order (default is FALSE).}

\item{wlabels}{logical value whether to write a CSV file with low/high (below/above the cutoff) group sample labels (default is TRUE).}

\item{wpdf}{logical value whether to write a PDF file with survival curves and plots (default is TRUE).}
}
\value{
no return value
}
\description{
Creates Kaplan-Meier survival curves for each feature from a validation data set by using a file with previously determined
stratification thresholds (one threshold per feature), and calculates the log-rank test p-value.
}
\examples{

Basic usage:
kmvalcut(input1="cutoffs.txt",input2="validation.txt",sfname="survival.txt",wdir="c:/test")

Example with built-in data files:

library(stats)
library(survival)
library(stringr)
library(data.table)
library(tools)
library(pracma)
library(kmcut)

# Load example gene expression data and survival data for 2 genes and 295 samples
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")
sdat = system.file("extdata", "survival_data_295.txt", package="kmcut")

# "example_genes_295_KM_quant_50.txt" is a file with cutoffs created by 'kmqcut'
# and must exist in directory "c:/test"
kmvalcut(input1="example_genes_295_KM_quant_50.txt",input2=fdat,sfname=sdat,wdir="c:/test")

This will create three output files in directory "c:/test":
1) PDF file with plots: "example_genes_295_KM_val.pdf"
2) Tab-delimited text file with the results: "example_genes_295_KM_val.txt"
3) CSV file with low/high sample labels: "example_genes_295_KM_val_labels.csv"
}
