% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposetable.R
\name{transposetable}
\alias{transposetable}
\title{Transpose a data table: convert rows to columns and columns to rows.
Row names will become column names, and column names will become row names.}
\usage{
transposetable(fnamein, fnameout, wdir)
}
\arguments{
\item{fnamein}{character vector that specifies the name of tab-delimited text file with the input data table.}

\item{fnameout}{character vector that specifies the name of output file where the transposed data table will be saved.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.}
}
\value{
no return value
}
\description{
Transpose a data table: convert rows to columns and columns to rows.
Row names will become column names, and column names will become row names.
}
\examples{

Basic usage:
transposetable(fnamein="table.txt", fnameout="transposedtable.txt", wdir="c:/test")

Example with built-in data files:

library(data.table)
library(kmcut)

# Load example gene expression data table for 2 genes
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")

transposetable(fnamein=fdat, fnameout="example_genes_295_transposed.txt", wdir="c:/test")

This will create a tab-delimited text file with the transposed table in directory "c:/test":
"example_genes_295_transposed.txt"
}
