% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucoxbatch.R
\name{ucoxbatch}
\alias{ucoxbatch}
\title{For each feature in the input file, this function performs a univariate Cox regression with the likelihood ratio test.}
\usage{
ucoxbatch(fname, sfname, wdir, min_uval = 50, psort = FALSE,
  verbose = TRUE)
}
\arguments{
\item{fname}{character vector that specifies the name of the file with feature(s) for each sample. The file must be tab-delimited,
where features are in rows and samples are in columns. First column must contain feature names. Column names must contain sample ids.}

\item{sfname}{character vector that specifies the name of the file with right-censored survival time data. The file must be tab-delimited,
where samples are in rows. First column must be named sample_id and contain sample ids that match those in fname. The file must
contain columns called stime and scens, with survival time and censoring variable (0 or 1), respectively.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.
Output file names are automatically created by adding "_ucoxbatch.txt" to 'fname'.}

\item{min_uval}{numeric value that specifies the minimal percentage of unique values per feature (default is 50)
Features that have less than 'min_uval' percent unique values are excluded from the analysis.}

\item{psort}{logical value whether to sort the output table by p-values in increasing order (default is FALSE).}

\item{verbose}{logical value whether to print progress (default is TRUE).}
}
\value{
no return value
}
\description{
For each feature in the input file, this function performs a univariate Cox regression with the likelihood ratio test.
}
\examples{

Basic usage:
ucoxbatch(fname="table.txt", sfname="survival.txt", wdir="c:/test")

Example with built-in data files:

library(stats)
library(survival)
library(stringr)
library(data.table)
library(tools)
library(pracma)
library(kmcut)

# Load example gene expression data and survival data for 2 genes and 295 samples
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")
sdat = system.file("extdata", "survival_data_295.txt", package="kmcut")

ucoxbatch(fname=fdat, sfname=sdat, wdir="c:/test")

This will create a tab-delimited text file with the results in directory "c:/test":
"example_genes_295_ucoxbatch.txt"
}
