% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucoxpred.R
\name{ucoxpred}
\alias{ucoxpred}
\title{For each feature in the input file, this function fits a univariate Cox regression model on a training dataset and then uses the model
to predict risk scores for a test dataset.}
\usage{
ucoxpred(fname1, sfname1, fname2, sfname2, wdir, min_uval = 50,
  psort = FALSE, verbose = TRUE)
}
\arguments{
\item{fname1}{character vector that specifies the name of the training file with feature(s) for each sample. The file must be tab-delimited,
where features are in rows and samples are in columns. First column must contain feature names. Column names must contain sample ids.}

\item{sfname1}{character vector that specifies the name of the file with right-censored survival time data for training dataset. The file must be tab-delimited,
where samples are in rows. First column must be named sample_id and contain sample ids that match those in fname1. The file must
contain columns called stime and scens, with survival time and censoring variable (0 or 1), respectively.}

\item{fname2}{character vector that specifies the name of the test file with feature(s) for each sample. The file must be tab-delimited,
where features are in rows and samples are in columns. First column must contain feature names. Column names must contain sample ids.}

\item{sfname2}{character vector that specifies the name of the file with right-censored survival time data for test dataset. The file must be tab-delimited,
where samples are in rows. First column must be named sample_id and contain sample ids that match those in fname2. The file must
contain columns called stime and scens, with survival time and censoring variable (0 or 1), respectively.}

\item{wdir}{character vector that specifies the name of the working directory for the input/output files.
Three output files are automatically created by adding:\cr'_cox_train_sum.txt' to 'fname1',
'_train_score.txt' to 'fname1';\cr'_test_score.txt' to 'fname2';}

\item{min_uval}{numeric value that specifies the minimal percentage of unique values per feature (default is 50).
Features that have less than 'min_uval' percent unique values are excluded from the analysis.}

\item{psort}{logical value whether to sort the output table by p-values in increasing order (default is FALSE).}

\item{verbose}{logical value whether to print progress (default is TRUE).}
}
\value{
no return value
}
\description{
For each feature in the input file, this function fits a univariate Cox regression model on a training dataset and then uses the model
to predict risk scores for a test dataset.
}
\examples{

Basic usage:
ucoxpred(fname1="train.txt", sfname1="survivaltrain.txt", fname2="test.txt", sfname2="survivaltest.txt", wdir="c:/test")

Example with built-in data files:

library(stats)
library(survival)
library(stringr)
library(data.table)
library(tools)
library(pracma)
library(kmcut)

# Load example gene expression data and survival data for 2 genes and 295 samples
fdat = system.file("extdata", "example_genes_295.txt", package="kmcut")
sdat = system.file("extdata", "survival_data_295.txt", package="kmcut")

# Use the example dataset as a training dataset and as a test dataset (perform resubstitution).
# NOTE: In a typical real world application training and test datasets will be different from one another.
ucoxpred(fname1=fdat, sfname1=sdat, fname2=fdat, sfname2=sdat, wdir="c:/test")

This will create three output files in directory "c:/test":
1) Tab-delimited text file with Cox summary for the training data: "example_genes_295_cox_train_sum.txt"
2) Tab-delimited text file with risk scores for training data: "example_genes_295_train_score.txt"
3) Tab-delimited text file with risk scores for test data: "example_genes_295_test_score.txt"
}
